+++ bitcoin/src/crypto/sighash.rs
@@ -1120,7 +1120,7 @@ mod tests {
     use crate::crypto::key::PublicKey;
     use crate::crypto::sighash::{LegacySighash, TapSighash};
     use crate::internal_macros::hex;
-    use crate::p2p::constants::Network;
+    use crate::network::Network;
     use crate::taproot::TapLeafHash;
 
     extern crate serde_json;
